+++ ReleaseNotes.md
@@ -1,3 +1,3 @@
-#Release Notes
+# Release Notes
 
-- Update SDK 15.5.2-beta-1
+- Update to Xamarin Android Plugin (linked with Android SDK 16.0.0-rc2)